requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
authlib==1.6.0
baml-py==0.90.2
bcrypt==4.3.0
black==25.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click>=8.0.0
cloudpickle==3.1.1
croniter==3.0.4
cryptography==45.0.4
exceptiongroup==1.3.0
fastapi==0.115.13
grpcio==1.69.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
isort==5.13.2
jinja2==3.1.6
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
multidict==6.5.0
mypy-extensions==1.1.0
openapi-pydantic==0.5.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
postgrest==0.16.2
prompt-toolkit==3.0.51
protobuf==4.25.8
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.0
pygments==2.19.2
pynacl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
rich==13.9.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sse-starlette==2.3.6
starlette==0.46.2
supabase==2.10.0
typeguard==2.13.3
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.34.3
watchdog==4.0.2
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
